pyproject.toml
[project]
name = "mma"
version = "1.0.2"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.3",
    "pygame_gui>=0.6.14",
    "Pillow>=10.0.0",
    "pytablericons>=1.0.0",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
